# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderOrderType } from "./orderOrderType";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  dailyNumber: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  items: OrderItem[];
  totalPrice: number;
  deliveryFee: number;
  discountCode?: string | null;
  discountAmount?: number | null;
  orderType: OrderOrderType;
  status: OrderStatus;
  paymentMethod: string;
  notes?: string | null;
  createdAt: string;
}